backend/requirements.txt
flask==3.1.0
flask-socketio==5.5.1
python-dotenv==1.1.0
openai==1.76.0
gunicorn==21.2.0
eventlet==0.35.2
